organize components folder, prefer kebab case, prefer dynamic path imports
