+++ __tests__/components/Search.spec.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 
-import Search from "@/components/Search";
+import Search from "@/components/boards/list/search";
 
 import {
   mockedQuery,
