+++ src/app/boards/edit/page.tsx
@@ -2,9 +2,12 @@
 
 import dynamic from "next/dynamic";
 
-const ImageEditor = dynamic(() => import("@/components/ImageEditor"), {
-  ssr: false,
-});
+const ImageEditor = dynamic(
+  () => import("@/components/boards/create/image-editor"),
+  {
+    ssr: false,
+  },
+);
 
 export default function Page() {
   return <ImageEditor />;
