+++ src/components/boards/detail/board-details.tsx
