+++ src/components/boards/detail/color-toggle.tsx
@@ -1,7 +1,6 @@
 import { cn } from "@/lib/utils";
-
-import { Label } from "./ui/label";
-import { Switch } from "./ui/switch";
+import { Label } from "@/components/ui/label";
+import { Switch } from "@/components/ui/switch";
 
 const ColorToggle = ({
   handleColorChange,
