+++ src/components/boards/detail/game-card.tsx
