+++ src/components/nav-bar-login-provider.tsx
@@ -1,5 +1,5 @@
 import { getCurrentSession } from "@/lib/session";
-import NavBar from "@/components/NavBar";
+import NavBar from "@/components/nav-bar";
 
 export default async function NavBarLoginProvider() {
   const { user } = await getCurrentSession();
